backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pdfplumber>=0.10.0
spacy>=3.7.0
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
xlsxwriter>=3.1.0
python-multipart
supabase>=2.0.0
